Replace automatic test execution by Travis CI with GitHub Actions
Travis CI has more or less limited its service to paying customers and
withdrawn from activities in open source communities.

Reintroduce test runs for all PHP versions 5.6 and later, as this package
is declared compatible with 5.6 and later.

Create test coverage reports only when needed, not by default, as Xdebug
may not be installed and needed to run the tests.