Correct handling of character sets in Syllable::hyphenateHtml()
The DOMDocument implementation uses the character set "ISO-8859-1" by
default. This led to incorrect encodings of UTF-8-specific characters
and entities, such as Cyrillic characters, dashes or the German
umlauts ä, ö, ü and ß, and thus to a partial failure of hyphenation
for these languages.

With this patch, the algorithm automatically detects the encoding of
the incoming HTML and inserts the corresponding meta tag
```
<meta http-equiv='content-type' content='text/html; charset=<encoding>'>
```
explicitly, which helps DOMDocument to use the correct internal
character encoding.

The algorithm of Syllable::hyphenateHtml() now

- extracts the body content from the incoming HTML,
- adds a simple HTML structure including the content-type meta tag,
- hyphenates the HTML,
- extracts the hyphenated body content and
- wraps the body content with the original HTML structure (if any).

This way, only the body content is hyphenated, and no description text
in the head, for example.

Furthermore, it is no longer necessary

- to explicitly tag-wrap the HTML to provide a single root tag,
- nor to remove unexpected DOMDocument internal default DTD or HTML
  structures from the hyphenated HTML at the end of processing,
- nor to explicitly set LIBXML_HTML_NOIMPLIED or LIBXML_HTML_NODEFDTD
  via Syllable::setLibxmlOptions(),
- nor to prepare the HTML with the expensive
  ```
  mb_convert_encoding($html, 'HTML-ENTITIES', "UTF-8")
  ```
  before passing it to Syllable::hyphenateHtml().